Fixed bot reloading problem
Bash had a weird behaviour when exec from the inside of the USR1 signal
handler. The bot could correctly exec himself, then rearming the signal
for the next reloading request, but this new signal was never caught. I
suppose it is because the bot somehow never really exited the previous
signal handler.

I will report this strange behaviour to the bash mailing list.